fix(edit aquaplant page)
